requirements.txt
Flask==2.3.3
tensorflow==2.18.0
tensorflow-probability==0.23.0
Pillow==9.2.0
numpy==2.0.2
scipy==1.13.1
gunicorn==20.1.0

